[backend] refactor: add folders in users
