+++ apps/portal-api/src/modules/organization-management/users/user-domain/users.domain.test.ts
@@ -20,7 +20,7 @@ import {
   loadUserConnection,
   updateUser,
   updateUserAtLogin,
-} from '../users.domain';
+} from '../user-domain/users.domain';
 
 //Issue with test
 describe('Users domain', () => {
