+++ apps/portal-api/src/modules/organization-management/users/users.resolver.test.ts
@@ -42,8 +42,12 @@ import {
   deleteSubscription,
   insertSubscription,
 } from '../../subcription/subscription.helper';
+import {
+  deleteUserById,
+  loadUser,
+  loadUserBy,
+} from './user-domain/users.domain';
 import { UserOrganizationPendingDomain } from './user-pending/user-organization-pending.domain';
-import { deleteUserById, loadUser, loadUserBy } from './users.domain';
 import { removeUser } from './users.helper';
 import usersResolver from './users.resolver';
