+++ apps/portal-api/src/modules/user_service/user-service.domain.test.ts
@@ -17,7 +17,7 @@ import UserService, {
 } from '../../model/kanel/public/UserService';
 import UserServiceCapability from '../../model/kanel/public/UserServiceCapability';
 import * as mailService from '../../server/mail-service';
-import { loadUserBy } from '../organization-management/users/users.domain';
+import { loadUserBy } from '../organization-management/users/user-domain/users.domain';
 import { removeUser } from '../organization-management/users/users.helper';
 import { createSubscription } from '../subcription/subscription.domain';
 import { SubscriptionStatus } from '../subscription.const';
